patch 2

+++ src/org/macesdev/stillflap/assets/lang/en_US.java
@@ -5,5 +5,5 @@ public class en_US {
 	public static String checkForUpdates_level2 = "Checking for new updates..";
 	public static String checkForUpdates_level3 = "Checking for new updates...";
 	public static String pressenter = "Press ENTER to Start Game";
-	public static float version = 1.0f;
+	public static double version = 1.0;
 }
